Separate match vocabulary components
Need to simplify the Match Vocabulary component, because the drag and drop
it's not working on the Modal window somehow.
